Remove idiotic unwrap() functions, scrubbed_arr
Both mlocker<T> and scrubbed<T> have ADL "unwrap" functions that simple
cast to the `T` base type.  This is idiotic and pointless because
casting to a public base type is already implicitly done in C++.

scrubbed_arr is idiotic in a different way: it just obfuscates the type
without actually providing a useful abstraction.