Make crypto object bool operators explicit
Avoid unintended conversions.

This caches one potentially nasty bug in particular where we are calling

    vector<string>.push_back({string, crypto::null_hash}

This happened to compile because crypto::hash was previously implicitly
convertible to `bool', and so this ended up calling
std::string's (string, size_t pos) substring constructor by implicitly
converting the null_hash to a bool (which was false, since it's null)
and then from there casting to a size_t, and ending up calling
(string, 0).  Since 0 is the start position of the whole string,
miraculously this didn't break anything.