Replace crypto:: randomness with libsodium
Remove Monero NIH RNG with libsodium's well-tested, more robust version.

Removes the terrible "thread-safe" random value generator which wasn't
actually properly thread safe at all, but merely slapped a mutex around
the random generator.

Removes the ability to "add extra entropy".  This seems like a
well-intentioned but terribly dangerous option: for the vast majority of
users this is likely to be no improvement (or make things worse), and
for the small number of users who know better and are on systems where
entropy is a problem, this is something better done with external tools
like haveged.

This also renames the random generation functions to have more
descriptive names:

- `crypto::rand<T>()` becomes `crypto::random_filled<T>()`.
- `crypto::rand_idx(...)` and `rand_range` become `random_index` and
  `random_range`.
- 'crypto::generate_random_bytes_thread_safe` becomes
  `crypto::fill_random` (and is now actually thread-safe via libsodium
  without needing a mutex on every call).
- `crypto::random_scalar` is gone, replaced with a direct call to
  libsodium's `crypto_core_ed25519_scalar_random`