Fix shared-ringdb disabling on testnet/devnet/fakenet
On mainnet, you can disable (by design) use of the shared-ringdb by
using `--shared-ringdb-dir` with an empty argument, but the code
handling the argument on testnet/devnet/fakenet appended
`testnet`/`devnet`/`fake` to the path, which made it non-empty and thus
failed to disable it as intended.

This fixes it.