Add `balance()` and `StoreFee()` helpers; formatting
`balance(x)` is a shortcut for `coins(x, x)`, which is particularly
useful in a bunch of places where `x` is a complex expression and so
currently we're doing messy things like `(coins(xxxxxxxxxxxx),) * 2`.

StoreFee() is a helper class for storing a fee, replacing the
`store_fee()` function that was heavily duplicated in the test code.