Handle reward rate of 0 in pool contract
Ethereum seems to return "0x" when there are no rewards. This breaks the
hex -> U64 conversion code because it uses `std::from_chars`. Giving an
empty string to `std::from_chars` sets the error code to a non-success
value causing the function to fail and throw.

This breaks the local-devnet without any logs. Here we handle that case
explicitly by checking the string size ourselves.