Make ETH transaction types consistent with smart contract
Oxen currently refers to the two ethereum service node leave events
using four different names: "deregister", "exit", "removal", and
"liquidated"; I've found this quite confusing to work with when trying
to compare with the smart contract because of the conflicting names,
especially "deregister" which has a quite different meaning in Oxen
("deregister" = SN kicked a node off for failing obligations).

This unifies the terminology to match with the smart contract emitted
events:

- removal (or remove) for a service node that has been removed
  (sometimes previously "exit")
- liquidated (or liquidation) for a liquidated service node (sometimes
  previously "deregister")
- removal request (instead of unlock or unlock request -- there was
  nothing particularly wrong with "unlock", but renaming unifies it with
  the smart contract terminology)

This commit is almost entirely just internal
function/method/struct/comment renames, but this does also affect the
OMQ interface for retrieving a signature.

Small related changes here that are more than just a simple
rename/reformatting:

- bls_aggregator: static functions were scattered all over; this groups
  them into an equivalent single anon namespace instead.
- got rid of the KEY_IMAGE_AWAITING_UNLOCK_HEIGHT = 0 constant; the
  "UNLOCK" term doesn't match eth removal requests anymore, but also
  it's clearer to simply write `if (requested_unlock_height)` rather
  than `if (requested_unlock_height !=
  KEY_IMAGE_AWAITING_UNLOCK_HEIGHT)`.
- added some more details around some comments about is_premature_unlock
  as to *why* it's safe to remove after HF20.  (It's safe because this
  is only used for mempool selection rather than consensus, and was
  added as a soft enforcement of the rule because the hard consensus
  rule originally added in HF19 meant to enforce no early small
  contributor unlocks erroneously applied at 3749 atomic OXEN instead of
  3749 OXEN).