Fix high_resolution_clock portability problem in removal signing
std::chrono::high_resolution_clock was being used for the removal
signing request timestamp, but that doesn't guarantee that it has a unix
time epoch (it typically does on unix systems, where it's an alias for
std::chrono::system_clock, but not on Windows where it's a
std::chrono::steady_clock), and so its use here (which we use for
extracting a timestamp) is a potential portability issue.  Switch to
std::chrono::system_clock, which is required to have a unix epoch.

It also appears that we had a mismatch between nanoseconds on the
message creator side (time_since_epoch().count()) but interpreting it as
seconds on the receiving side.

There was also a (less important) endian issue with sending the
timestamp as the uint64_t bytes, which is easily fixed by
sending/receiving it as a plain integer instead.