Remove dead liquidated tx code
We don't (and apparently never have) actually used the
ServiceNodeLiquidated event: we don't pick it up during log scanning,
and so there has never been a way for a liquidation tx to actually go
out on the network because we never translate them.

Moreover we don't *need* it: the contract emits both a
`ServiceNodeLiquidated` *and* a `ServiceNodeRemoval` during liquidation,
so just handling the ServiceNodeRemoval already handles both regular +
liquidated removals.