Fix: always serialize reward in HF19
This was broken from a half-edited implementation in the previous
commits: we still use and want the serialized `reward` in HF19, it's
just height and winner that should be under the Oxen 10 compat mode.
(We still have it in HF20, too, but in HF20+ it's serialized in
`block_header` instead of `block`).

This fixes it by moving the reward serialization outside the compat
branches.