Fix Sepolia errors fetching recent reward heights
On Sepolia nets, we sometimes get an error trying to retrieve reward
rates, particularly at startup when we are performing initial fetches.

It turns out this is because Arbitrum Sepolia often goes a bit slower
than 250ms (currently average is more like 360ms) as 250ms is only a
minimum block time, but Arb will sometimes slow down a bit if there are
no blocks when the sequencer makes the next block.

But compounding this is the fact that most Arb nodes only keep 128
blocks + blocks up to 30m old, and so requesting the reward at a height
>30m old will error, which can happen in the current implementation when
the L2 block rate slows down.

Even on mainnet with its very-close-to-250ms block time average the
warning would trigger *occassionally* because if you are starting up at
current L2 height of 1002399 then the heights you need rewards for are
1000000, 997600, 995200, but 995200 is 7199 ago and so is right at the
30min deletion threshold.

This includes a couple fixes to address it:

- Only fetch current and last applicable block reward.  We were
  previously also fetching the second-last one, but with the last one
  being at least 10 minutes old, this is unnecessary as we should never
  need it.  (And so for mainnet, in the worst case example above, we
  would only ever fetch back to 997600, which will typically be only 20
  minutes ago).
- Lower the reward interval on testnet/devnet/stagenet to 600 blocks
  instead of mainnet's 2400 blocks.

With the two changes above, we should be fine for a sustained average L2
block rate (over the last half hour) of up to around 1.5s on Sepolia,
and up to 375ms on mainnet Arbitrum.

It's also worth noting that hitting these isn't fatal: if the L2 chain
stalls for some reason there is an additional buffer provided by Oxen
nodes that have been running for since before the stall: they will hold
onto cached recent values and can still proceed (only recently restarted
ones would have difficulty validating).

Also removes the L2_BLOCK_TIME because it really isn't quite correct
(and was only being used in one place).