Fix mempool selector
The mempool meta info was not setting l2_height for non-L2 event
transactions, which left l2_height as an uninitialized value, but this
would later break when selecting transactions because such a transaction
would have an L2 height that failed the max-L2-height check with:

    [txpool:debug|cryptonote_core/tx_pool.cpp:1917] Considering cf7d58354caf7e8c62fe37a2cd84268d139ee5f0195f04b5f156d286c9fffe75, weight 472, current block weight 0/299400, current reward 0.000000000
    [txpool:debug|cryptonote_core/tx_pool.cpp:1925]   state change from L2 height 94205658036688 is not in in admissable L2 heights <= 73294606

which caused non-L2 events to get stuck in the mempool until some node
happened to lead a pulse quorum that had a sufficiently low value.

This fixes it to zero-initialize the meta struct so that this won't
happen (and then removes all the now-redundant zero assignments).