Sync staking requirement from SN rewards contract
This adds support for oxend propagating a staking requirement change
from the staking contract into oxend, via the existing L2 event
confirmation mechanism.

Staking requirement is now retrieved via the service_node_list state,
rather than a free `get_staking_requirement` function as it now depends
on state changes and confirmations.  The old free function is renamed to
`get_default_staking_requirement` to make it clear that it isn't
(necessarily) the correct staking requirement value anymore.

This also drops the `height` input field from the
`get_staking_requirement` RPC endpoint; it was only really there for
when the staking requirement was height dependent, which it hasn't been
for a long time, and won't work with the change here for historical
values because we don't have a state_t for every possible height.