prepare_eth_registration: remove multi-contributor address
This extra address is no longer used (and wasn't really usable anyway
because there can't be a contribution contract until after you run this
command).

The POP now always signs the operator address, and so this drops the
extra address both from the command, and from submission to the staking
website.