Break up load blocks into subsystem into sections
- Store hardfork heights at the top of the function to reduce line
length and minimise the amount of line-breaking enforced by the
formatter. This breaks up the text-heavy noise in the height enumeration step.

- Separate the step where we calculate the minimum height of all the
subsystems which is the height to start loading blocks from to after
enumerating all the heights. This makes it very clear what heights we
need to run the min() function on and find the earliest height.

- Remove the vector accumulating heights. It's clearer now what heights to min()
now that the enumerating height and processing the height is separated.

- Avoid signed int math. Unsigned works fine by relying on well-defined
wrapping behaviour which removes the noisy static_casts in the main loop.

- Remove early exit checking on there being no blocks to check. The
bottom of the function had the exact same check. The loop is well
behaved for (total_blocks == 0) so we can just let the function proceed
rather than add another branch.