Rewrite state serialization to execute linearly
- Short term serialisation was broken because the height that was
calculated as the cutoff height to serialise was incorrectly adding a
offset to the current height before being subtracted again such that all
the SNL states that were serialised only serialised their quorums.

This means that every time we reorg, 1 or 2 blocks, this triggers a full
rescan from the last 10k-th block interval. This was hidden in the code
because the logic was hard to read and modify.

- The serialisation routine for some reason was also trying to serialise
the quorums at a cut-off height, but, when you process a block into the
SNL you _already_ cull history that is too old. Hence when you go to
serialise, the list of states has already been pruned, there's no need
to have logic to prune it again. All we have to do is serialise it to
disk. That simplifies the serialisation routine which unveiled the bug
in the above bulletpoint.

- Remove tricky logic that migrated the short-term window into the
archive window. This had a bunch of checks embedded in a loop with some
arithmetic to calculate some offsets from the height. This is not
necessary if we just serialise the state on block add by determining if
the current state falls under the criteria to be serialised into the
archive or not. This simplifies the loop, de-nests the algorithm and
hence reduces the cognitive load required to grok what's happening.

- Fix an issue where SNL::load() calls reset(false). Reset was clearing
out the SQL DB causing a full-rescan on startup every time even if the
SNL and SQL DB's were in sync.