Per-network reg URLs; rpc cleanup; output modes
Makes the default staking URL per-network.  Currently both stagenet and
mainnet are pointing at stake.getsession.org, but stagenet will move
somewhere else once mainnet is ready.  The others don't have defaults
currently and will error if you run the command expecting one.

Renames and re-documents the bls_registration_request endpoint to
contract_registration.  It was oddly misnamed because the `_request`
suffix implies it makes a request (like BLS_EXIT_LIQUIDATION_REQUEST),
but it does not, it simply obtains from the local node.  This also
renames some of the arguments/parameters to better align with the names
we're settled on in oxend and the L2 contract.

Adds plaintext and contract `register` print modes for advanced uses:
- `register 0xWhatever plaintext` prints an unformatted, plaintext
  version of `print`, which is easier to use if extracting values via a
  script.
- `register 0xWhatever contract` prints the values split up/formatted
  for easier use when directly interacting with a smart contract, such
  as:

      register 0xabcdefabcdefabcdefabcdefabcdefabcdefabcd contract

      L2 Contract Registration Information for 1b671bb0e49c12ad2f47b86581da87f3bbec24d97bc0f10645a986f47649acd1:
      Operator address: 0xABcdEFABcdEFabcdEfAbCdefabcdeFABcDEFabCD
      ServiceNodeRewards.addBLSPublicKey()/multi-contract.reset...() parameters:
         blsPubkey/key=(0x128e9a0d5cb5fb4c72ab476632c97f8e9bdb71b6c537ca6790b3b11ea3c01ab6, 0x201601a84d3bf8d8f05a0a5256e4e7d760b744791c8257b9333b0ed83047d8ee)
         blsSignature/sig=(0x2cf6604b93b2eb5be1f40169d569ac469a7492b5a1ec2eb325b34edc47c78414, 0x1cae5ff894923fb6c8e6403370c96e107700265cfd1cba746139a222e1b6f351, 0x1f2931de75023747eacbc1df52293354285e6b5761988c448de62e7052ec8718, 0x1a23744ceae498528e820ed249ca81b6b9aab0804e6c0bbdd26227a6e978eab7)
         serviceNodeParams/params=(0x1b671bb0e49c12ad2f47b86581da87f3bbec24d97bc0f10645a986f47649acd1, 0x039db33dc750ee15eddaaa2e035fbdecb281dc37838e4ef1080cb90570a16203, 0x1edb3ba03e9ce179fc70f8cc318c18d96ed4eb0cd03840c1ac4350c1743b1c06, 0)