Allow detach of SNL and SQL to lower height than requested
We detach to a height that is supported by both SNL and SQL by checking
the backups available on both before proceeding. This ensures in
exceptional cases if the SQL side is missing a backup, the SNL doesn't
go ahead and rewind itself. This would cause a desync that can only be
remedied by noticing it, and then restarting the daemon.

Instead the 2 double check what is a safe height they both support and
do that patching up that particular avenue of breakage.