Remove subtract_sn_rewards from db_sqlite
We don't support popping a block from the SQL DB anymore, only detach to
a recent height which is supported by the SNL, ONS and SQL DB. This is
required for deterministic behaviour on rewind. Since that's the case we
can remove the subtract code from the DB and simplify.