Store delayed payment history, fix delayed payments being applied too early
Delayed payments with a delay blocks of 1 was being applied to the SQL
DB before its SQL height was updated. The delayed payment was being committed at
the current height of the SQL DB (which was behind by one block as it
had not yet processed the block yet).

The block would be added to the DB updating the height of the SQL DB and
cause the triggers to execute and remove the delayed payments scheduled
for the current height of the SQL DB (e.g. the new one we just added)
before it can be paid out to the user in question.