C++20 & updated lib compilation (DROP FROM DEV!)
Current stable does not build with update libraries that now require
C++20; this copies some of the fixes back from dev to make C++20 work.

NOTE TO FUTURE oxen-core MAINTAINERS: when attempt to merge Oxen 11 dev
into stable to make a stable Oxen 11.1.0 release, this commit will
conflict and can just be dropped entirely: there is *nothing* included
here that does not already have the same or a superior implementation in
dev.