BLS: check funded SNs instead of active SNs
Decommed nodes (i.e. funded but inactive) are still potentially
available for signing off on BLS requests, but we were only considering
active SNs for making connections for BLS sigs.

This also splits up/genericizes the different modes of the
`is_service_node()` call to remove the optional bool changing the
meaning of the function, and adds a new `is_funded_service_node()` for
what we need here.