Fix pre-HF19 subsystem rescan resetting to 0
If a subsystem rescan is interrupted (e.g. by hitting Ctrl-C) before
reaching HF19, the are-the-dbs-in-sync check fail because SQLite doesn't
have any payment rows, which ends up completely resetting the service
node list and triggering a full rescan from 101250.

This commit fixes it by simply allowing any pre-HF19 height regardless
of what is in sqlite, since "empty" is the correct state for such a
case (and the SQLite's blockchain_detached will delete anything that
might be there, making it correctly empty).