Redo config handling without ucf
This drops ucf handling of the config file and instead just takes a more
conservative approach to editing the files:

- If a service-node-public-ip=, l2-oxend=, or l2-provider= is already
  present in the oxen.conf then postinst will not try to do anything
  with the file, and config won't even try asking since we won't need
  the value.

  This fixes the issue where people who already set up an l2 provider or
  proxy with earlier 11.1.x releases were still getting asked for this
  info, which is pointless.  We were trying to replace it, but that's
  error prone and annoying.

This means the public-ip and l2 info questions should only be asked the
first time they are needed, and after that its up to the admin to edit
the config files.

This also makes the config file handling more conservative so that we
can drop ucf handling, which often caused trouble, especially if someone
elected *not* to install the package maintainers version of the config
file.