Move systemd "ready" notification earlier in the startup
Currently we only send READY=1 to systemd after any block popping or
blockchain rescanning, but unfortunately those can take an extended
period, during which the `systemctl start oxend` command will be stalled
waiting for it.

This moves it to just before we start popping or rescanning.  This is a
bit unfortunate that we won't (from a systemctl perspective) fail to
start for various runtime reasons (such as being unable to bind to the
p2p or omq ports), but that seems better than making "systemctl start
oxend" stall for an hour (most notably including on upgrades).