HF22: use rounding instead of truncation for rewards
This should average out across all accounts over time to 0, but even if
it doesn't, it's never going to be worse than truncation, and there's
enough slack in the block reward (e.g. from oxen-side block reward
adjustment) that it isn't going to run over what the rewards contract
has in it.