Fix cursed² issues in sqlite tests for BLOB addresses
Cursed chaingen code was selecting as string and reinserting as string,
which combined sqlite allowing you to insert any type into any column
(and simply ignoring the column type) was resulting in
batched_payment_accrued (in core_tests) having apparent primary key
violations where a binary address was present as both a TEXT and as a
BLOB which aren't equal.

This updates the unit test code to properly select and insert blobs.