Remove unnecessary const ref on Pulse's OMQ thread ID
The thread ID is designed to be safe and cheap to copy (it's internally
just an int, or, 4 bytes on a majority modern platforms). This member is
used in different threads (i.e. the OMQ worker threads) and reads are
atomic so it also helps with a potential race condition in certain
contexts (like the local devnet).