HF23: Disallow SN deregs that skip decomm
In theory a misbehaving obligations quorum could submit a direct
deregistration even when the SN being tested should transition (or
should remain) decommission rather then getting deregistered, however
this was not actually being enforced on chain.

This fixes it for HF23+ so that such invalid deregs are not valid in a
block so that even if a rogue operator gets lucky and gains control of
an obligations quorum (or manipulates pulse entropy to gain such
control), they won't actually be able to do any worse than
decommissioning good nodes, which would then simply get recommissioned
by honest nodes.