Fix BLS pubkey validation in uptime proof checks
The current code does not verify that the BLS pubkey we received in an
uptime proof actually matches the BLS pubkey that was used in
registration: rather it simply checks that *some* BLS key was specified,
and that the included pop matches that BLS key.

This fixes it.

This will affect a handful of mainnet nodes at the next hardfork (either
11.7 or 12.0), which are currently sending such invalid signatures.